requirements.txt
amqp==2.6.1
appdirs==1.4.3
asgiref==3.3.4
autopep8==1.5.6
billiard==3.6.4.0
boto3==1.17.109
botocore==1.20.109
CacheControl==0.12.6
celery==4.4.7
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cloudinary==1.25.0
colorama==0.4.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==3.2
django-braces==1.14.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-oauth-toolkit==1.5.0
django-rest-framework-social-oauth2==1.1.0
django-rest-knox==4.1.0
django-rest-passwordreset==1.1.0
django-storages==1.11.1
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-gis==0.17
drf-extra-fields==3.1.1
drf-social-oauth2==1.0.9
drf-yasg==1.20.0
Faker==8.4.0
Flask==2.0.1
flower==0.9.7
geographiclib==1.52
geopy==2.2.0
geos==0.2.3
gunicorn==20.1.0
h11==0.12.0
html5lib==1.0.1
httpcore==0.13.2
httpx==0.18.1
humanize==3.5.0
idna==2.8
importlib-metadata==4.0.0
inflection==0.5.1
install==1.3.4
ipaddr==2.2.0
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
jwcrypto==0.8
kombu==4.6.11
lockfile==0.12.2
lxml==4.6.3
MarkupSafe==2.0.1
msgpack==0.6.2
oauth2-provider==0.0
oauthlib==3.1.0
packaging==20.3
paystack==1.5.0
paystackapi==2.0.0
pep517==0.8.2
Pillow==8.2.0
progress==1.5
prometheus-client==0.8.0
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.6
python-crontab==2.5.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytoml==0.1.21
pytz==2021.1
redis==3.5.3
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rfc3986==1.4.0
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
s3transfer==0.4.2
simplejson==3.17.2
six==1.14.0
sniffio==1.2.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tornado==6.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
whitenoise==5.3.0
zipp==3.4.1
